Toast params & better restoring & translations for lost daily

+++ src/hooks/useAppUpdateIfNeeded.ts
@@ -4,10 +4,13 @@ import getAppStatus from '@api/getAppStatus';
 
 export default function useAppUpdateIfNeeded( ) {
     useEffect(() => {
-        getAppStatus().then(({ shouldUpdate }) => {
+        getAppStatus().then(({ shouldUpdate, shouldResetStore }) => {
             if (shouldUpdate) {
-                // This will be controled in future versions
-                localStorage.clear();
+                // This will be more controlled in the future
+                if (shouldResetStore) {
+                    localStorage.clear();
+                }
+
                 location.reload();
             }
         });
